Improved speed
Figured out that i built Kernels every time that kernel is used. Much better performance after fixing it.

+++ c_implementation/weights/network1/b1.h
@@ -1,3 +1,3 @@
 //Testiranje ispisa
 //Niz je: (512,)
-extern double b1[512];
+extern float b1[512];
