+++ c_implementation/weights/network1/b2.h
@@ -1,3 +1,3 @@
 //Testiranje ispisa
 //Niz je: (256,)
-extern double b2[256];
+extern float b2[256];
