+++ c_implementation/weights/network1/b3.h
@@ -1,3 +1,3 @@
 //Testiranje ispisa
 //Niz je: (128,)
-extern double b3[128];
+extern float b3[128];
