+++ c_implementation/weights/network1/w1.h
@@ -1,3 +1,3 @@
 //Testiranje ispisa
 //Niz je: (784, 512)
-extern double w1[512][784];
+extern float w1[512][784];
