+++ c_implementation/weights/network1/w2.h
@@ -1,3 +1,3 @@
 //Testiranje ispisa
 //Niz je: (512, 256)
-extern double w2[256][512];
+extern float w2[256][512];
