+++ c_implementation/weights/network1/w3.h
@@ -1,3 +1,3 @@
 //Testiranje ispisa
 //Niz je: (256, 128)
-extern double w3[128][256];
+extern float w3[128][256];
