+++ c_implementation/weights/network1/wout.h
@@ -1,3 +1,3 @@
 //Testiranje ispisa
 //Niz je: (128, 10)
-extern double wout[10][128];
+extern float wout[10][128];
