+++ c_implementation/weights/network2/layer1.h
@@ -1,3 +1,3 @@
 //Testiranje ispisa
 //Niz je: (2048,)
-extern double layer1[2048];
+extern float layer1[2048];
